fixes na extração de metricas da coleção
