+++ opac_proc/extractors/ex_collections.py
@@ -1,4 +1,5 @@
 # coding: utf-8
+import requests
 from datetime import datetime
 from opac_proc.datastore.models import ExtractCollection
 from opac_proc.extractors.base import BaseExtractor
@@ -11,6 +12,8 @@ if config.DEBUG:
     logger = getMongoLogger(__name__, "DEBUG", "extract")
 else:
     logger = getMongoLogger(__name__, "INFO", "extract")
+
+PUBLICATION_SIZE_ENDPOINT = 'ajx/publication/size'
 
 
 class CollectionExtactor(BaseExtractor):
@@ -25,6 +28,51 @@ class CollectionExtactor(BaseExtractor):
         self.get_instance_query = {
             'acronym': self.acronym
         }
+
+    def _get_json_metrics(self, metric_name, url, params):
+        json_data = {}
+
+        try:
+            response = requests.get(url, params=params)
+        except Exception, e:
+            logger.error(u'Erro recuperando as metricas: %s (url=%s, msg=%s)' % (metric_name, url, str(e)))
+        else:
+            if response.status_code == 200:
+                json_data = response.json()
+        return json_data
+
+    def _extract_metrics(self):
+        params = {
+            'code': self.acronym,
+            'collection': self.acronym,
+        }
+        # references:
+        url = '{0}/{1}'.format(config.OPAC_METRICS_URL, PUBLICATION_SIZE_ENDPOINT)
+        params['field'] = 'citations'
+        references = self._get_json_metrics('references', url, params)
+
+        # articles:
+        url = '{0}/{1}'.format(config.OPAC_METRICS_URL, PUBLICATION_SIZE_ENDPOINT)
+        params['field'] = 'documents'
+        articles = self._get_json_metrics('articles', url, params)
+
+        # issues:
+        params['field'] = 'issue'
+        url = '{0}/{1}'.format(config.OPAC_METRICS_URL, PUBLICATION_SIZE_ENDPOINT)
+        issues = self._get_json_metrics('issues', url, params)
+
+        # jornals:
+        params['field'] = 'issn'
+        url = '{0}/{1}'.format(config.OPAC_METRICS_URL, PUBLICATION_SIZE_ENDPOINT)
+        journals = self._get_json_metrics('jornals', url, params)
+
+        metrics = {
+            'total_citation': int(references.get('total', 0)),
+            'total_article': int(articles.get('total', 0)),
+            'total_issue': int(issues.get('total', 0)),
+            'total_journal': int(journals.get('total', 0)),
+        }
+        return metrics
 
     @update_metadata
     def extract(self):
@@ -68,4 +116,8 @@ class CollectionExtactor(BaseExtractor):
         else:
             # atualizo self.metadata para que self.children_ids seja salvo junto com self.raw_data no save()
             self.metadata['children_ids'] = self.children_ids
+
+        # extração de metricas
+        self._raw_data['metrics'] = self._extract_metrics()
+
         logger.info(u'Fim CollectionExtactor.extract(%s) %s' % (self.acronym, datetime.now()))
